requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
nltk==3.8.1
openai==0.27.9
requests==2.31.0
newspaper3k==0.2.8